Have working distance and duration calculations

+++ db/migrate/20141210220540_create_restaurants.rb
@@ -5,7 +5,7 @@ class CreateRestaurants < ActiveRecord::Migration
       t.string :address
       t.string :cuisine
       t.integer :cost
-      t.integer :rating
+      t.decimal :rating
 
       t.timestamps
     end
